clippy: replace the cast allows with checked conversions
cast_possible_truncation/wrap/sign_loss were allowed crate-wide. Convert
offsets with a saturating u64->usize helper, so an out-of-range value
(only reachable on 32-bit) fails the bounds check instead of wrapping;
convert allocation sizes with a checked helper that errors instead of
issuing a huge request; widen cstr to a u64 offset. The string-table,
symbol, and ld-cache writes narrow through checked u32/u16. The d_tag
and ELF32-pointer truncations are intentional and keep a local allow.

+++ tests/mutating_ops.rs
@@ -467,7 +467,7 @@ fn rename_dynamic_symbols_matches_reference() {
         .filter_map(|i| {
             let b = &img.section_data[dynsym][i * 24..];
             let st_name = u32::from_le_bytes([b[0], b[1], b[2], b[3]]);
-            formatelf::ir::cstr(&img.section_data[dynstr], st_name)
+            formatelf::ir::cstr(&img.section_data[dynstr], u64::from(st_name))
         })
         .find(|n| n.bytes().next().is_some_and(|c| c.is_ascii_alphabetic()))
         .unwrap()
